Struct UAV_aggregator::map::Map

pub struct Map {
    _walls: Obj,
    _min: Vector3<f32>,
    _max: Vector3<f32>,
    _step: Vector3<f32>,
    facesInChunk: HashMap<Vector3<usize>, HashSet<Face>>,
    pub collisionPlusEps: f32,
    pub collisionMinusEps: f32,
    pub COR: f32,
    pub mi_s: f32,
    pub mi_d: f32,
    pub minimalDist: f32,
}

Simulation map

Fields
_walls: Obj_min: Vector3<f32>_max: Vector3<f32>_step: Vector3<f32>facesInChunk: HashMap<Vector3<usize>, HashSet<Face>>collisionPlusEps: f32collisionMinusEps: f32COR: f32mi_s: f32mi_d: f32minimalDist: f32
Implementations

impl Map

pub fn new( path: &str, collisionPlusEps: f32, collisionMinusEps: f32, grid: Vector3<f32>, COR: f32, mi_s: f32, mi_d: f32, minimalDist: f32 ) -> Self

Constructor

pub fn checkWalls(&self, point: Vector3<f32>, radius: f32) -> Vec<Vector3<f32>>

Checks if in specified point there is collision with map walls. Returns colection of normal vectors of face that point colide with. If there is no collisions, return colection length is equal 0.

pub fn checkWalls2( &self, start_point: Vector3<f32>, velocity: Vector3<f32>, dt: f32, radius: f32 ) -> Vec<Vector3<f32>>

Modified checkWalls, predict position in looptime

pub fn checkWallsBest(&self, point: Vector3<f32>) -> Option<(f32, Vector3<f32>)>

Checks if in specified point there is collision with map walls. Returns normal vector of wall that is the closest to point If there is no collisions, return None WARNING: this function may work incorrectly when face has not got true normal vector

pub fn checkWallsBest2( &self, point: Vector3<f32>, velocity: Vector3<f32>, dt: f32 ) -> Option<(f32, Vector3<f32>)>

Checks if in specified point there will collide with map walls in specificed dt. Returns normal vector of wall that will be cross in next dt If there is no collisions, return None

fn calcChunk(&self, point: Vector3<f32>) -> Vector3<usize>

Calculates chunk that for specified point

fn insertFace(&mut self)

Splits faces into chunks

pub fn getMinMax(&self) -> (Vector3<f32>, Vector3<f32>)

Get minimal and maximal coordinate of cubiod that contains all map’s vertices