Struct UAV_aggregator::objects::Objects

pub struct Objects {
    _ctx: Context,
    pub _time: Arc<Mutex<f32>>,
    pub states: Arc<Mutex<Vec<ObjectState>>>,
    pub info: Arc<Mutex<HashMap<usize, ObjectInfo>>>,
    running: Arc<AtomicBool>,
    control_socket: Socket,
    _state_proxy: Option<JoinHandle<()>>,
    _state_cupturer: Option<JoinHandle<()>>,
    _dropListener: Option<(JoinHandle<()>, JoinHandle<()>)>,
}

All objects in simulation

Fields
_ctx: Context_time: Arc<Mutex<f32>>states: Arc<Mutex<Vec<ObjectState>>>info: Arc<Mutex<HashMap<usize, ObjectInfo>>>running: Arc<AtomicBool>control_socket: Socket_state_proxy: Option<JoinHandle<()>>_state_cupturer: Option<JoinHandle<()>>_dropListener: Option<(JoinHandle<()>, JoinHandle<()>)>
Implementations

impl Objects

pub fn new(_ctx: Context, port: usize) -> Self

Constructor

fn parseInfo( time: &Arc<Mutex<f32>>, states: &Arc<Mutex<Vec<ObjectState>>>, info: String )

Parses objects state from string

fn _sendControlMsg(&self, msg: &str) -> String

Sends control message to object’s simulation

pub fn addObj( &self, mass: f32, CS: f32, pos: Vector3<f32>, vel: Vector3<f32>, obj_info: ObjectInfo ) -> isize

Add new object to simulation

pub fn removeObj(&self, id: usize)

Remove object from simulation

pub fn updateWind(&self, wind: Vec<(usize, Vector3<f32>)>)

Peridically updates winds info for objects

pub fn setForce(&self, id: usize, force: Vector3<f32>)

Sets outer force value applied to object specified by id

pub fn sendSurfaceCollison( &self, id: usize, COR: f32, mi_s: f32, mi_d: f32, normalVector: &Vector3<f32> )

Sends information about collision with surface to object specified by id

pub fn getPositions(&self) -> Vec<(usize, Vector3<f32>)>

Get position of all objects in air

pub fn getVelocities(&self) -> Vec<(usize, Vector3<f32>)>

Get velocities of all objects in air

pub fn getPosVels(&self) -> Vec<(usize, Vector3<f32>, Vector3<f32>)>

Get positions & velocities of all objects in air

pub fn getPosVelsRadius(&self) -> Vec<(usize, Vector3<f32>, Vector3<f32>, f32)>

Get positions, velocities & radius of collision of all objects in air